refactor: migrate widgets code
